Rewritten MIDI patch table to accommodate velocity-switched SoundFont
Due to severe memory leaks in the Alchemy-compiled synthesizer currently in use, we were forced to implement a custom patching scheme as a workaround.

The synthesizer was struggling to correctly address multiple patches without leaking enormous amounts of memory. To mitigate this, we created a velocity-switched SoundFont file, where instrument patches are accessed by sending `noteOn` messages with specific velocity values.

Despite attempting to disable the default "Velocity to Attenuation" modulator in the SoundFont file, the synthesizer still applied unwanted velocity-based volume scaling. To counteract this behavior, we arranged instruments in a way that compensates for their natural dynamics:
- Softer instruments were mapped to higher velocity values.
- Louder instruments were mapped to lower velocity values.